Backend/api/requirements.txt
agno==1.1.9
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.16
Flask==3.1.0
flask-cors==5.0.1
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
gotrue==2.11.4
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.2.0
openai==1.66.0
packaging==24.2
postgrest==0.19.3
propcache==0.3.0
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
PyMuPDF==1.25.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
PyJWT==2.10.1
realtime==2.4.1
rich==13.9.4
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.14.0
supafunc==0.9.3
tomli==2.2.1
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
websockets==14.2
Werkzeug==3.1.3
yarl==1.18.3
